+++ fe/src/tpl-bootstrap/components/Document.vue
@@ -43,6 +43,9 @@
 </template>
 
 <script>
+/* Lightbox for Bootstrap 5 */
+import Lightbox from 'bs5-lightbox'
+
 import {markdownRender} from "@/_shared/utils/docms_utils"
 import {useRoute} from 'vue-router'
 import {watch} from 'vue'
@@ -71,6 +74,7 @@ export default {
   },
   computed: {
     documentContentRendered() {
+      this._updateLightbox()
       return markdownRender(this.$localedText(this.document.content), {
         sanitize: true,
         tags: this.$siteMeta.tags,
@@ -78,6 +82,11 @@ export default {
     },
   },
   methods: {
+    _updateLightbox() {
+      this.$nextTick(() => {
+        document.querySelectorAll('[data-toggle="lightbox"]').forEach(el => el.addEventListener('click', Lightbox.initialize));
+      })
+    },
     _fetchSiteMeta(vue) {
       vue.$fetchSiteMeta(
           () => vue.status = 0,